backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.26.0

# AI & LangChain Ecosystem (The likely conflict source)
langchain>=0.1.0
langchain-core
langchain-google-genai
langgraph
google-generativeai

# Database
supabase>=2.3.0